Lab4

+++ Lab3/Homework/Lab3/src/main/java/com/example/Lab3/models/Customer.java
@@ -1,17 +1,17 @@
 package com.example.Lab3.models;
 
 public class Customer {
-    private final Long id;
+    private final int id;
     private final String name;
     private final boolean loyal;
 
-    public Customer(Long id, String name, boolean loyal) {
+    public Customer(int id, String name, boolean loyal) {
         this.id = id;
         this.name = name;
         this.loyal = loyal;
     }
 
-    public Long getId() { return id; }
+    public int getId() { return id; }
     public String getName() { return name; }
     public boolean isLoyal() { return loyal; }
 }
