+++ Lab3/Homework/Lab3/src/main/java/com/example/Lab3/models/Order.java
@@ -3,17 +3,17 @@ package com.example.Lab3.models;
 import java.math.BigDecimal;
 
 public class Order {
-    private final Long id;
+    private final int id;
     private final Long customerId;
     private final BigDecimal totalAmount;
 
-    public Order(Long id, Long customerId, BigDecimal totalAmount) {
+    public Order(int id, Long customerId, BigDecimal totalAmount) {
         this.id = id;
         this.customerId = customerId;
         this.totalAmount = totalAmount;
     }
 
-    public Long getId() { return id; }
+    public int getId() { return id; }
     public Long getCustomerId() { return customerId; }
     public BigDecimal getTotalAmount() { return totalAmount; }
 }
